añadidos filtros al get de detail

+++ src/main/java/org/example/apitiendaaa/domain/DTO/ProductOutDTO.java
@@ -11,5 +11,5 @@ public class ProductOutDTO {
     private long id;
     private String name;
     private float price;
-    private float categoryId;
+    private long categoryId;
 }
